가맹점 수정 완료

+++ src/main/java/com/boot/ict05_final_admin/domain/store/dto/StoreDetailDTO.java
@@ -58,7 +58,7 @@ public class StoreDetailDTO {
     private StoreType storeType;
 
     /** 매장 총 직원수 */
-    private int storeTotalEmployees;
+    private Long storeTotalEmployees;
 
     /** 계약 시작일 */
     private LocalDate storeContractStartDate;
